requirements.txt
PyQt5
paho-mqtt==1.5.1
requests
Flask
win11toast
python-dotenv